upload: set hw_rng_model during creation
We're hitting this weird issue where sometimes, the call to `update()`
gives:

```
glanceclient.exc.HTTPException: 415 Unsupported Media Type
The request media type application/octet-stream is not supported by this
server.
```

Which I think is because the client is using the wrong MIME type for the
JSON patch operation. We should be able to set the property from the
outset during the `create()` call, so let's do that.